# projects/web/Jolteon/src/jolteon/services/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spark, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/models/spell/list": {
      "get": {
        "tags": [
          "SpellModelController"
        ],
        "parameters": [
          {
            "name": "list",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ISpellModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/spell/filtered": {
      "get": {
        "tags": [
          "SpellModelController"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ISpellModelFilterDefinition"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ISpellModelFilterDefinition"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ISpellModelFilterDefinition"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ISpellModelFilterDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ISpellModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/spell/byString/{str}": {
      "get": {
        "tags": [
          "SpellModelController"
        ],
        "parameters": [
          {
            "name": "str",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ISpellModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/spell/{id}": {
      "get": {
        "tags": [
          "SpellModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ISpellModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SpellModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SpellModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpellModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ISpellModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SpellModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/stats/filtered": {
      "get": {
        "tags": [
          "StatsModelController"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/IStatsFilterDefinition"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IStatsFilterDefinition"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IStatsFilterDefinition"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IStatsFilterDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IStats"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/stats/{id}": {
      "get": {
        "tags": [
          "StatsModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IStats"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "StatsModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/IStats"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IStats"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IStats"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IStats"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IStats"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatsModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/status/filtered": {
      "get": {
        "tags": [
          "StatusModelController"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/IStatusModelFilterDefinition"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IStatusModelFilterDefinition"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IStatusModelFilterDefinition"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IStatusModelFilterDefinition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IStatusModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/status/byString/{str}": {
      "get": {
        "tags": [
          "StatusModelController"
        ],
        "parameters": [
          {
            "name": "str",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IStatusModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/status/{id}": {
      "get": {
        "tags": [
          "StatusModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IStatusModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "StatusModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/StatusModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StatusModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IStatusModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatusModelController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActorType": {
        "enum": [
          "Source = 0",
          "Target = 1"
        ],
        "type": "string",
        "format": ""
      },
...
      "CharacteristicIdIStatEntityDictionary": {
        "type": "object",
        "properties": {
          "entityUid": {
            "type": "string"
          },
          "dic": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/IStat"
            }
          }
        },
        "additionalProperties": false
      },
      "CharacteristicIdIStatKeyValuePair": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/IStat"
          }
        },
        "additionalProperties": false
      },
      "CharacteristicIdMathEquationEntityDictionary": {
        "type": "object",
        "properties": {
          "entityUid": {
            "type": "string"
          },
          "dic": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/MathEquation"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ConditionComparatorType": {
        "enum": [
          "EQ = 0",
          "NE = 1",
          "GT = 2",
          "GE = 3",
          "LT = 4",
          "LE = 5"
        ],
        "type": "string",
        "format": ""
      },
      "ConditionGroupType": {
        "enum": [
          "AND = 0",
          "OR = 1"
        ],
        "type": "string",
        "format": ""
      },
...
      "DeleteResult": {
        "type": "object",
        "properties": {
          "deletedCount": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "isAcknowledged": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "Direction9Type": {
        "enum": [
          "center = 0",
          "top = 1",
          "topright = 2",
          "right = 3",
          "bottomright = 4",
          "bottom = 5",
          "bottomleft = 6",
          "left = 7",
          "topleft = 8"
        ],
        "type": "string",
        "format": ""
      },
...
      "ICondition": {
        "type": "object",
        "properties": {
          "actorType": {
            "$ref": "#/components/schemas/ActorType"
          },
          "comparator": {
            "$ref": "#/components/schemas/ConditionComparatorType"
          },
          "groupType": {
            "$ref": "#/components/schemas/ConditionGroupType"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ICondition"
            },
            "nullable": true
          }
        },
        "additionalProperties": {
          "type": "object"
        }
      },
...
      "ISpellModel": {
        "type": "object",
        "properties": {
          "nameId": {
            "type": "string"
          },
          "descriptionId": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "skinIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "sourceCondition": {
            "$ref": "#/components/schemas/ICondition"
          },
          "targetFilter": {
            "$ref": "#/components/schemas/ICondition"
          },
          "statsId": {
            "type": "string"
          },
          "rangeZoneMin": {
            "$ref": "#/components/schemas/IZone"
          },
          "rangeZoneMax": {
            "$ref": "#/components/schemas/IZone"
          },
          "modelUid": {
            "type": "string",
            "nullable": true
          },
          "entityUid": {
            "type": "string"
          },
          "effectIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ISpellModelFilterDefinition": {
        "type": "object",
        "additionalProperties": false
      },
...
      "IStat": {
        "type": "object",
        "properties": {
          "statId": {
            "type": "string"
          },
          "entityUid": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "IStats": {
        "type": "object",
        "properties": {
          "base": {
            "$ref": "#/components/schemas/CharacteristicIdIStatEntityDictionary"
          },
          "growth": {
            "$ref": "#/components/schemas/CharacteristicIdMathEquationEntityDictionary"
          },
          "entityUid": {
            "type": "string"
          },
          "keys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IStat"
            },
            "nullable": true,
            "readOnly": true
          },
          "pairs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacteristicIdIStatKeyValuePair"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IStatsFilterDefinition": {
        "type": "object",
        "additionalProperties": false
      },
...
      "IStatusModel": {
        "type": "object",
        "properties": {
          "nameId": {
            "type": "string"
          },
          "descriptionId": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "skinIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "statsId": {
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/StatusPriorityType"
          },
          "modelUid": {
            "type": "string",
            "nullable": true
          },
          "entityUid": {
            "type": "string"
          },
          "effectIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IStatusModelFilterDefinition": {
        "type": "object",
        "additionalProperties": false
      },
...
      "IVector2": {
        "type": "object",
        "properties": {
          "x": {
            "type": "integer",
            "format": "int32"
          },
          "z": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "IVector3": {
        "type": "object",
        "properties": {
          "x": {
            "type": "integer",
            "format": "int32"
          },
          "z": {
            "type": "integer",
            "format": "int32"
          },
          "y": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "IZone": {
        "type": "object",
        "properties": {
          "zoneType": {
            "$ref": "#/components/schemas/ZoneType"
          },
          "size": {
            "$ref": "#/components/schemas/IVector3"
          },
          "negative": {
            "type": "boolean"
          },
          "worldOrigin": {
            "$ref": "#/components/schemas/ActorType"
          },
          "worldOffset": {
            "$ref": "#/components/schemas/IVector2"
          },
          "localOrigin": {
            "$ref": "#/components/schemas/Direction9Type"
          },
          "rotation": {
            "$ref": "#/components/schemas/Rotation4Type"
          },
          "canRotate": {
            "type": "boolean",
            "format": "boolean",
            "nullable": true
          },
          "sizeIndexExtendFromSource": {
            "type": "integer",
            "format": "int32"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IZone"
            },
            "nullable": true
          },
          "maxSampleCount": {
            "type": "integer",
            "format": "int32"
          },
          "samplingType": {
            "$ref": "#/components/schemas/TargetSamplingType"
          }
        },
        "additionalProperties": false
      },
...
      "MathEquation": {
        "type": "object",
        "properties": {
          "functions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MathFunction"
            }
          }
        },
        "additionalProperties": false
      },
      "MathFunction": {
        "type": "object",
        "properties": {
          "xFromIncluded": {
            "type": "integer",
            "format": "int32"
          },
          "xToExcluded": {
            "type": "integer",
            "format": "int32"
          },
          "slopes": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "additionalProperties": false
      },
...
      "Rotation4Type": {
        "enum": [
          "top = 0",
          "right = 1",
          "bottom = 2",
          "left = 3"
        ],
        "type": "string",
        "format": ""
      },
...
      "SpellModel": {
        "type": "object",
        "properties": {
          "entityUid": {
            "type": "string"
          },
          "modelUid": {
            "type": "string",
            "nullable": true
          },
          "nameId": {
            "type": "string"
          },
          "descriptionId": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "skinIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "sourceCondition": {
            "$ref": "#/components/schemas/ICondition"
          },
          "targetFilter": {
            "$ref": "#/components/schemas/ICondition"
          },
          "statsId": {
            "type": "string"
          },
          "effectIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "rangeZoneMin": {
            "$ref": "#/components/schemas/IZone"
          },
          "rangeZoneMax": {
            "$ref": "#/components/schemas/IZone"
          }
        },
        "additionalProperties": false
      },
...
      "StatusModel": {
        "type": "object",
        "properties": {
          "entityUid": {
            "type": "string"
          },
          "modelUid": {
            "type": "string",
            "nullable": true
          },
          "nameId": {
            "type": "string"
          },
          "descriptionId": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "skinIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "statsId": {
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/StatusPriorityType"
          },
          "effectIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StatusPriorityType": {
        "enum": [
          "System = 0",
          "Passive = 1",
          "Status = 2"
        ],
        "type": "string",
        "format": ""
      },
...
      "TargetSamplingType": {
        "enum": [
          "closestToOrigin = 0",
          "furthestToOrigin = 1",
          "closestToSource = 2",
          "furthestToSource = 3",
          "random = 4"
        ],
        "type": "string",
        "format": ""
      },
...
      "ZoneType": {
        "enum": [
          "point = 0",
          "line = 1",
          "diagonal = 2",
          "cross = 3",
          "xcross = 4",
          "star = 5",
          "crossHalf = 6",
          "xcrossHalf = 7",
          "circle = 8",
          "circleHalf = 9",
          "square = 10",
          "squareHalf = 11",
          "rectangle = 12",
          "ellipse = 13",
          "ellipseHalf = 14",
          "board = 15",
          "offboard = 16",
          "all = 17"
        ],
        "type": "string",
        "format": ""
      }
    }
  }
}